requirements.txt
requests==2.22.0
Flask==1.1.2
Flask_RESTful==0.3.8
Pillow==8.1.2
PyPDF2==1.26.0
